add some files

+++ include/data_stream.h
@@ -14,9 +14,9 @@
 
 namespace kdtree{
     std::vector<uint32_t> read_index(std::string file_name);
-    std::vector<double> read_val(std::string file_name);
+    std::vector<float> read_val(std::string file_name);
 
-    std::vector<std::vector<double>> read_data(std::string row_file_name, 
+    std::vector<std::vector<float>> read_data(std::string row_file_name, 
             std::string col_file_name, std::string data_file_name);
 
     void write(int32_t * buf, std::string file_name, uint32_t size);
